:wrench: 调整控制摇杆消息的数据类型
将 `control_joy_message_t` 结构体中的 `x` 和 `y` 字段的数据类型从无符号的 `uint8_t` 更改为有符号的 `int8_t`。这一调整允许摇杆数据能够表示负值，从而支持反向控制。

具体变更如下：
- **固件 (`robot-embedded-firmware`)**：
    - 更新 `control_joy_message_t` 和 `Wrobot` 结构体，将摇杆坐标 `x` 和 `y` 的类型从 `uint8_t` 和 `int` 分别改为 `int8_t`。
    - 相应地，反序列化函数 `deserialize_control_joy_message` 从读取 `u8` 修改为读取 `i8`。

- **安卓应用 (`robot-android-app`)**：
    - 在 `ControlJoy.java` 中，移除 `PercentageValue.percentage` 的转换逻辑，直接将输入的 `int` 强制转换为 `byte`，以匹配固件的数据类型变更。

+++ robot-embedded-firmware/include/robot.hpp
@@ -50,10 +50,10 @@ typedef enum {
 } RobotState;
 
 typedef struct {
-  int joyy;
-  int joyy_last;
-  int joyx;
-  int joyx_last;
+  int8_t joyy;
+  int8_t joyy_last;
+  int8_t joyx;
+  int8_t joyx_last;
 } Wrobot;
 
 // typedef struct {
